Adding "CV Programming" mode
In normal mode, any attached CV's override the outputs.
An additional mode is added where CV's act just like manual inputs and can "pick up" the programmed values and set them.
In programmed mode, the dial is green if the incoming CV has "picked up" the preset value. In this case the output will be changed by altering the CV. If it's blue, the CV has not picked up the value and changing the CV will not automatically change the output.